third_party/blink/web_tests/html5lib/generated/run-tests7-write-expected.txt

CONSOLE ERROR: line 1: Uncaught SyntaxError: Unexpected token '<'
../resources/tests7.dat: PASS
